+++ foundation-api/src/main/java/dev/soffa/foundation/commons/validation/ValidationResult.java
@@ -1,0 +1,11 @@
+package dev.soffa.foundation.commons.validation;
+
+import java.util.Map;
+
+public interface ValidationResult {
+
+    Map<String,String> getErrors();
+
+    boolean hasErrors();
+
+}