+++ foundation-api/src/main/java/dev/soffa/foundation/core/Broadcast.java
@@ -1,0 +1,4 @@
+package dev.soffa.foundation.core;
+
+public interface Broadcast {
+}