+++ foundation-api/src/main/java/dev/soffa/foundation/core/Command.java
@@ -1,4 +1,4 @@
 package dev.soffa.foundation.core;
 
-public interface Command<I,O> extends Operation<I,O> {
+public interface Command<I, O> extends Operation<I, O>, Recorded {
 }