+++ foundation-api/src/main/java/dev/soffa/foundation/core/EventHandler.java
@@ -3,8 +3,8 @@ package dev.soffa.foundation.core;
 import dev.soffa.foundation.context.Context;
 import org.checkerframework.checker.nullness.qual.NonNull;
 
-public interface EventHandler<I> {
+public interface EventHandler<I, O> extends Operation<I,O>, Recorded {
 
-    void handle(I input, @NonNull Context ctx);
+    O handle(I input, @NonNull Context ctx);
 
 }