+++ foundation-api/src/main/java/dev/soffa/foundation/core/Operation.java
@@ -1,5 +1,6 @@
 package dev.soffa.foundation.core;
 
+import dev.soffa.foundation.commons.validation.ValidationResult;
 import dev.soffa.foundation.context.Context;
 import org.checkerframework.checker.nullness.qual.NonNull;
 
@@ -9,8 +10,8 @@ public interface Operation<I, O> {
 
     O handle(I input, @NonNull Context ctx);
 
-    default void validate(I input, @NonNull Context ctx) {
-        // do nothing
+    default ValidationResult validate(I input, @NonNull Context ctx) {
+        return null;
     }
 
 }