+++ foundation-starter/src/main/java/dev/soffa/foundation/spring/service/DefaultMessageHandler.java
@@ -1,13 +1,13 @@
 package dev.soffa.foundation.spring.service;
 
 import dev.soffa.foundation.commons.Logger;
+import dev.soffa.foundation.commons.Sentry;
 import dev.soffa.foundation.commons.TextUtil;
 import dev.soffa.foundation.config.OperationsMapping;
 import dev.soffa.foundation.context.Context;
 import dev.soffa.foundation.context.ContextHolder;
 import dev.soffa.foundation.core.Dispatcher;
 import dev.soffa.foundation.core.Operation;
-import dev.soffa.foundation.error.TechnicalException;
 import dev.soffa.foundation.message.Message;
 import dev.soffa.foundation.message.MessageFactory;
 import dev.soffa.foundation.message.MessageHandler;
@@ -40,28 +40,32 @@ public class DefaultMessageHandler implements MessageHandler {
             dispatcher.set(this.context.getBean(Dispatcher.class));
         }
 
-        final Context context = message.getContext();
-        ContextHolder.set(context);
-        TenantHolder.set(context.getTenantId());
+
         Object operation = mapping.getInternal().get(message.getOperation());
         if (operation == null) {
             LOG.debug("Message %s skipped, no local handler registered", message.getOperation());
             return Optional.empty();
         }
 
+        final Context context = message.getContext();
+        ContextHolder.set(context);
+        TenantHolder.set(context.getTenantId());
+
         if (authManager != null && context.hasAuthorization()) {
             authManager.handle(context);
         }
 
-        LOG.debug("New message received with operation %s#%s", message.getOperation(), message.getId());
+        LOG.info("[pubsub] New message received with operation %s#%s", message.getOperation(), message.getId());
 
         if (!(operation instanceof Operation)) {
-            throw new TechnicalException("Unsupported operation type: " + operation.getClass().getName());
+            Logger.app.error("[pubsub] unsupported operation type: " + operation.getClass().getName());
+            return Optional.empty();
         }
 
         Class<?> inputType = mapping.getInputTypes().get(message.getOperation());
         if (inputType == null) {
-            throw new TechnicalException("Unable to find input type for operation " + message.getOperation());
+            Logger.app.error("[pubsub] Unable to find input type for operation " + message.getOperation());
+            return Optional.empty();
         }
 
         final AtomicReference<Object> payload = new AtomicReference<>();
@@ -69,10 +73,10 @@ public class DefaultMessageHandler implements MessageHandler {
         if (message.getPayload() != null) {
             if (TextUtil.isNotEmpty(message.getPayloadType())) {
                 try {
-                    LOG.debug("Deserializing message content into %s", message.getPayloadType());
+                    LOG.debug("[pubsub] Deserializing message content into %s", message.getPayloadType());
                     payload.set(MessageFactory.getPayload(message, Class.forName(message.getPayloadType())));
                 } catch (ClassNotFoundException e) {
-                    LOG.error("Unable to deserialize message into %s", message.getPayloadType());
+                    LOG.error("[pubsub] Unable to deserialize message into %s", message.getPayloadType());
                 }
             }
             if (payload.get() == null) {
@@ -86,16 +90,19 @@ public class DefaultMessageHandler implements MessageHandler {
         } else {
             LOG.debug("Invoking operation %s with payload of type %s", operation.getClass().getSimpleName(), payload.get().getClass().getSimpleName());
         }
-        TenantHolder.set(context.getTenantId());
-        //noinspection unchecked
-        Object result = dispatcher.get().invoke((Operation<Object, Object>) operation, payload.get(), context);
-        if (result == null) {
-            return Optional.empty();
-        }
-        if (result instanceof ResponseEntity) {
-            //TODO: handle status ?
-            result = ((ResponseEntity<?>) result).getData();
-        }
-        return Optional.of(result);
+
+        return Sentry.get().watch("pubsub event: " + message.getOperation(), () -> {
+            @SuppressWarnings("unchecked")
+            Object result = dispatcher.get().invoke((Operation<Object, Object>) operation, payload.get(), context);
+
+            if (result == null) {
+                return Optional.empty();
+            }
+            if (result instanceof ResponseEntity) {
+                //TODO: handle status ?
+                result = ((ResponseEntity<?>) result).getData();
+            }
+            return Optional.of(result);
+        });
     }
 }